requirements.txt
flask
firebase-admin
google-cloud-dialogflow
gunicorn
python-dotenv
pandas